remove HomeFeed, moved logic to HomePage
